initial commit
Incomplete.

- reusable popup overlay UI element and JS wiring.
- dialog-based HWI initial support.
- HWI.enumerate workaround to call installed CLI
