+++ templates/base.html
@@ -6,6 +6,7 @@
 	<link rel="stylesheet" type="text/css" href="/static/styles.css?{{rand}}">
 </head>
 <body>
+{% include "includes/overlay.html" %}
 <div class="row holder">
 	<nav class="side">
 		<a href="/settings/" class="item core">
